+++ src/lib/index.js
@@ -1,7 +1,5 @@
 /* eslint-disable global-require */
 
-'use strict'
-
 module.exports = {
   api: require('./api'),
   es: require('./es'),