+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Correct content types are now returned
 - Searching for a nonexistent collection returns empty results
 - Re-ingesting an Item maintains the existing value of properties.created and sets properties.updated to now
+- Fixes the responses from `/collections/{collectionId}`, `/collections/{collectionId}/items`, and `/collection/{collectionId}/items/{itemId}`.
 
 ### Changed